fix: consolidate VersionUtil, add migration log warning, and clean imports
