+++ app/src/main/resources/io/apicurio/registry/storage/impl/sql/db-version
@@ -1,1 +1,2 @@
 108
+
